Rename wrapper_error.rs
- renare wrapper_error.rs to errors.rs in the bureaucracy module
